+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/util/TriConsumer.java
@@ -1,8 +1,10 @@
 package uk.ac.bbsrc.tgac.miso.core.util;
 
+import java.io.IOException;
+
 @FunctionalInterface
 public interface TriConsumer<A, B, C> {
 
-  public void accept(A a, B b, C c);
+  public void accept(A a, B b, C c) throws IOException;
 
 }